------------------------------------------- buffer optimization & OPEN-MP development ------------------------------------------- (last in r705)
* sequential TYPHON (Typhon-seq) renamed with $(opt) compilation option
-> Typhon-seq is linked to last compiled executable (for NRG cases)

* (VARCOM) buffer/packet functions split to PACKET module
. cell_buffer, face_buffer
. routine to compute index

* CPU results
(on a RK2 MUSCL-fast HLLC dambreak case)
buffer   fenrir(X5680)   mjollnir(E5440)
  64         316             502
 128         300             483
 256         294             480
 512         300             480
1024         315             502
2048         309             498
-> PACKET/cell_buffer is set to 256 (old value was 64)
   (+5% performance are expected)

* OPEN-MP directives in
. MONOTONIC post-limitation        (postlimit_monotonic)
. KDIF (Heat transfer) integration (integration_kdif_ust)
. Field x=y/x=a.x/x=x+y/x=x+a.y    (CFDTOOLS/BASEFIELD) 
. face buffer in residuals         (flux_to_res)
. NS source terms                  (calc_source_ext)
. NS time step                     (calc_ns_timestep)
-> speed-up increase (RK2 HLLC 1st order + source terms)
   2 threads: 1.54 to 1.71
   4 threads: 2.12 to 2.67
   8 threads: 2.48 to 3.75
   
NON-REGRESSION
. Ok

+++ SOURCE/EQNS/calc_source_ext.f90
@@ -7,8 +7,7 @@
 !------------------------------------------------------------------------------!
 subroutine calc_source_ext(umesh, field, defns, curtime)
 
-use TYPHMAKE
-use VARCOM
+use PACKET
 use OUTPUT
 use USTMESH
 use DEFFIELD
@@ -31,34 +30,53 @@ integer   :: ic	! index on cells
 real(krp) :: mass
 type(v3d) :: dmomentum, massV
 real(krp) :: dpower
+integer, pointer      :: ista(:), iend(:) ! starting and ending index
+integer               :: ib, buf, nblock      ! buffer size 
 
 ! -- BODY --
 
-call new_fct_env(blank_env)      ! temporary environment from FCT_EVAL
-call fct_env_set_real(blank_env, "t", curtime)
+if ((defns%is_extpower).or.(defns%is_extforce)) then
+  call new_buf_index(umesh%ncell_int, cell_buffer, nblock, ista, iend)
+  call new_fct_env(blank_env)      ! temporary environment from FCT_EVAL
+  call fct_env_set_real(blank_env, "t", curtime)
+endif
 
 !! DEV : planned to add x, y, z position 
 
 if (defns%is_extpower) then
-  do ic = 1, umesh%ncell_int
-    call fct_eval_real(blank_env, defns%extpower, dpower)
-    field%residu%tabscal(2)%scal(ic) = field%residu%tabscal(2)%scal(ic) + umesh%mesh%volume(ic,1,1) * dpower
+  !$OMP PARALLEL DO private(ic, dpower) shared(blank_env)
+  do ib = 1, nblock
+    buf = iend(ib)-ista(ib)+1
+    do ic = ista(ib), iend(ib)
+      call fct_eval_real(blank_env, defns%extpower, dpower)
+      field%residu%tabscal(2)%scal(ic) = field%residu%tabscal(2)%scal(ic) + umesh%mesh%volume(ic,1,1) * dpower
+    enddo
   enddo
+  !$OMP END PARALLEL DO
 endif
 
 if (defns%is_extforce) then
-  do ic = 1, umesh%ncell_int
-    call fct_eval_real(blank_env, defns%extforce_x, dmomentum%x)
-    call fct_eval_real(blank_env, defns%extforce_y, dmomentum%y)
-    call fct_eval_real(blank_env, defns%extforce_z, dmomentum%z)
-    mass   = umesh%mesh%volume(ic,1,1) * field%etatcons%tabscal(1)%scal(ic)
-    massV  = umesh%mesh%volume(ic,1,1) * field%etatcons%tabvect(1)%vect(ic)
-    field%residu%tabvect(1)%vect(ic) = field%residu%tabvect(1)%vect(ic) + (mass*dmomentum)
-    field%residu%tabscal(2)%scal(ic) = field%residu%tabscal(2)%scal(ic) + (massV.scal.dmomentum)
+  !$OMP PARALLEL DO private(ic, mass, massV, dmomentum) shared(blank_env)
+  do ib = 1, nblock
+    buf = iend(ib)-ista(ib)+1
+    do ic = ista(ib), iend(ib)
+      call fct_eval_real(blank_env, defns%extforce_x, dmomentum%x)
+      call fct_eval_real(blank_env, defns%extforce_y, dmomentum%y)
+      call fct_eval_real(blank_env, defns%extforce_z, dmomentum%z)
+      mass   = umesh%mesh%volume(ic,1,1) * field%etatcons%tabscal(1)%scal(ic)
+      massV  = umesh%mesh%volume(ic,1,1) * field%etatcons%tabvect(1)%vect(ic)
+      field%residu%tabvect(1)%vect(ic) = field%residu%tabvect(1)%vect(ic) + (mass*dmomentum)
+      field%residu%tabscal(2)%scal(ic) = field%residu%tabscal(2)%scal(ic) + (massV.scal.dmomentum)
+    enddo
   enddo
+  !$OMP END PARALLEL DO
 endif
 
-call delete_fct_env(blank_env)      ! temporary environment from FCT_EVAL
+if ((defns%is_extpower).or.(defns%is_extforce)) then
+  deallocate(ista, iend)
+  call delete_fct_env(blank_env)      ! temporary environment from FCT_EVAL
+endif
+
 
 end subroutine calc_source_ext
 !------------------------------------------------------------------------------!
